chapter3_4_aianimals/search/similar_image_search/proxy/pyproject.toml
[tool.poetry]
name = "proxy"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.10.6"
httpx = "^0.23.0"
Pillow = "^9.0.1"
numpy = "^1.22.3"
fastapi = "^0.75.0"
gunicorn = "^20.1.0"
uvicorn = "^0.17.6"
redis = "^4.1.4"
psycopg2-binary = "^2.9.3"
pydantic = "^1.9.0"
dependency-injector = "^4.39.1"
prometheus-client = "^0.14.1"
prometheus-fastapi-instrumentator = "^5.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
